+++ tests/integration/test_models/test_full_battery_models/test_lithium_ion/test_compare_basic_models.py
@@ -13,14 +13,17 @@ class TestCompareBasicModels:
         dfn = pybamm.lithium_ion.DFN()
 
         # Solve basic DFN
-        basic_sim = pybamm.Simulation(basic_dfn, parameter_values=parameter_values)
+        solver = pybamm.IDAKLUSolver(atol=1e-8, rtol=1e-8)
+        basic_sim = pybamm.Simulation(
+            basic_dfn, parameter_values=parameter_values, solver=solver
+        )
         t_eval = [0, 3600]
         t_interp = np.linspace(0, 3600)
         basic_sim.solve(t_eval=t_eval, t_interp=t_interp)
         basic_sol = basic_sim.solution
 
         # Solve main DFN
-        sim = pybamm.Simulation(dfn, parameter_values=parameter_values)
+        sim = pybamm.Simulation(dfn, parameter_values=parameter_values, solver=solver)
         sim.solve(t_eval=t_eval, t_interp=t_interp)
         sol = sim.solution
 
@@ -43,7 +46,10 @@ class TestCompareBasicModels:
         parameter_values = pybamm.ParameterValues("Chen2020_composite")
 
         # Solve basic DFN
-        basic_sim = pybamm.Simulation(basic_dfn, parameter_values=parameter_values)
+        solver = pybamm.IDAKLUSolver(atol=1e-8, rtol=1e-8)
+        basic_sim = pybamm.Simulation(
+            basic_dfn, parameter_values=parameter_values, solver=solver
+        )
         t_eval = [0, 3600]
         t_interp = np.linspace(0, 3600)
         basic_sim.solve(t_eval=t_eval, t_interp=t_interp)
@@ -69,14 +75,17 @@ class TestCompareBasicModels:
         spm = pybamm.lithium_ion.SPM()
 
         # Solve basic SPM
-        basic_sim = pybamm.Simulation(basic_spm, parameter_values=parameter_values)
+        solver = pybamm.IDAKLUSolver(atol=1e-8, rtol=1e-8)
+        basic_sim = pybamm.Simulation(
+            basic_spm, parameter_values=parameter_values, solver=solver
+        )
         t_eval = [0, 3600]
         t_interp = np.linspace(0, 3600)
         basic_sim.solve(t_eval=t_eval, t_interp=t_interp)
         basic_sol = basic_sim.solution
 
         # Solve main SPM
-        sim = pybamm.Simulation(spm, parameter_values=parameter_values)
+        sim = pybamm.Simulation(spm, parameter_values=parameter_values, solver=solver)
         sim.solve(t_eval=t_eval, t_interp=t_interp)
         sol = sim.solution